Update default order in read() to be '' (empty string)
read() should not impose a defined ordering to the user. There are places
where read() is called without allowing the user to change the default ordering
(e.g. in the SQL parser). By using '' as the default ordering, we preserve
whatever ordering the user has, rather than forcing it to 'start,end' (which
are not always desirable).